refactor: conditional rendering in recoveryview
