+++ ios/rehat Watch App/RecoveryView.swift
@@ -1,51 +1,52 @@
 import SwiftUI
 
-//struct Therapy: Identifiable, Decodable {
-//  var id: UUID?
-//  var name: String
-//  var slides: [Slide]
-//  var icon: String
-//  var color: String
-//  
-//  struct Slide: Decodable {
-//    var title: String
-//    var subTitle: String
-//  }
-//}
+struct Therapy: Identifiable, Decodable {
+  var id: UUID?
+  var name: String
+  var slides: [Slide]
+  var icon: String
+  var color: String
+  
+  struct Slide: Decodable {
+    var title: String
+    var subTitle: String
+  }
+}
 
-//class ReadDataTherapy: ObservableObject {
-//  @Published var therapies = [Therapy]()
-//  
-//  init(){
-//    loadData()
-//  }
-//  
-//  func loadData() {
-//    guard let url = Bundle.main.url(forResource: "recoveryData", withExtension: "json") else {
-//      print("JSON file not found")
-//      return
-//    }
-//    
-//    do {
-//      let data = try Data(contentsOf: url)
-//      var decodedTherapies = try JSONDecoder().decode([Therapy].self, from: data)
-//      
-//      for index in decodedTherapies.indices {
-//        if decodedTherapies[index].id == nil {
-//          decodedTherapies[index].id = UUID()
-//        }
-//      }
-//      
-//      self.therapies = decodedTherapies
-////      print("JSON data successfully loaded.")
-//    } catch let error {
-//      print("Error loading JSON data: \(error)")
-//    }
-//  }
-//}
+class ReadDataTherapy: ObservableObject {
+  @Published var therapies = [Therapy]()
+  
+  init(){
+    loadData()
+  }
+  
+  func loadData() {
+    guard let url = Bundle.main.url(forResource: "recoveryData", withExtension: "json") else {
+      print("JSON file not found")
+      return
+    }
+    
+    do {
+      let data = try Data(contentsOf: url)
+      var decodedTherapies = try JSONDecoder().decode([Therapy].self, from: data)
+      
+      for index in decodedTherapies.indices {
+        if decodedTherapies[index].id == nil {
+          decodedTherapies[index].id = UUID()
+        }
+      }
+      
+      self.therapies = decodedTherapies
+//      print("JSON data successfully loaded.")
+    } catch let error {
+      print("Error loading JSON data: \(error)")
+    }
+  }
+}
 
 struct RecoveryView: View {
-//  @EnvironmentObject var workoutManager: WorkoutManager
+  @EnvironmentObject var workoutManager: WorkoutManager
+  @EnvironmentObject var appState: AppState
   @ObservedObject var recoveryDatas = ReadDataTherapy()
   @ObservedObject var rnConnector: RNConnector
   @State private var isPresented: Bool = false
@@ -82,7 +83,7 @@ struct RecoveryView: View {
               }
             }
           } else if reference.label == "Grounding Technique" {
-            Section {
+            let grounding_techniques = {
               ForEach(recoveryDatas.therapies, id: \.id) { therapy in
                 NavigationLink(destination: DetailView(therapy: therapy)) {
                   VStack(alignment: .leading) {
@@ -97,47 +98,52 @@ struct RecoveryView: View {
                   }
                 }
               }
-            } header : {
-              Text("Grounding Technique")
+            }
+            
+            if appState.isPanic {
+              grounding_techniques()
+            } else {
+              Section {
+                grounding_techniques()
+              } header : {
+                Text("Grounding Technique")
+              }
             }
           }
-//          else if reference.label == "Emergency Call" {
-//            NavigationLink(destination: EmergencyContactsView(rnConnector: rnConnector)) {
-//              VStack(alignment: .leading) {
-//                Image(systemName: "phone")
-//                  .foregroundColor(.orange) // placeholder
-//                  .frame(width: 30, height: 30)
-//                  .padding(.top, 5)
-//                  .scaleEffect(1.5)
-//                Text(reference.label)
-//                  .padding(.bottom, 10)
-//                  .font(.title3)
-//              }
-//            }
-//          }
+          else if appState.isPanic && reference.label == "Emergency Call" {
+            NavigationLink(destination: EmergencyContactsView(rnConnector: rnConnector)) {
+              VStack(alignment: .leading) {
+                Image(systemName: "phone")
+                  .foregroundColor(.orange) // placeholder
+                  .frame(width: 30, height: 30)
+                  .padding(.top, 5)
+                  .scaleEffect(1.5)
+                Text(reference.label)
+                  .padding(.bottom, 10)
+                  .font(.title3)
+              }
+            }
+          }
         }
         .navigationDestination(isPresented: $isPresented, destination: {BreathView()})
       }
       .navigationTitle("Recovery")
-//      .containerBackground(.orange.gradient, for: .navigation)
     } detail: {
-      
       BreathView()
-      
     }
   }
 }
 
-//func colorFromString(_ name: String) -> Color {
-//    switch name {
-//    case "green":
-//        return Color.green
-//    case "blue":
-//        return Color.blue
-//    default:
-//        return Color.black
-//    }
-//}
+func colorFromString(_ name: String) -> Color {
+    switch name {
+    case "green":
+        return Color.green
+    case "blue":
+        return Color.blue
+    default:
+        return Color.black
+    }
+}
 
 #Preview {
   let rnConnector = RNConnector() // Create an instance of RNConnector
